requirements.txt
# WEEX AI Trading Hackathon - Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
ccxt>=4.5.0
